+++ package.json
@@ -59,7 +59,9 @@
     "@actions/core": "^1.2.7",
     "@actions/github": "^4.0.0",
     "@crowdin/cli": "^3.7.10",
+    "@docusaurus/module-type-aliases": "^2.2.0",
     "@docusaurus/types": "^2.1.0",
+    "@tsconfig/docusaurus": "^1.0.6",
     "@types/jest": "^26.0.23",
     "@types/unist": "^2.0.3",
     "del": "^6.0.0",
@@ -82,6 +84,8 @@
     "remark": "^13.0.0",
     "remark-gfm": "^1.0.0",
     "tar-stream": "^2.2.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.3",
     "unist-util-visit-parents": "^3.1.1"
   },
   "resolutions": {
